Added Working Offline / Online Delay
In our environment I noticed that the script sets up the account
properly but upon first launch doesn’t actually start syncing until you
close and reopen the app.  I believe the configuration is happening so
fast that Outlook gets caught in a weird race condition and it doesn’t
know that it needs to sync at the end.  So I added a section to throw
Outlook into “offline” mode, then added a 5 second delay at the end of
the config.  After that I throw it back into Online mode.  With these
changes this script has become a lot more consistent in our
environment.